+++ src/main/kotlin/site/siredvin/turtlematic/tags/BlockTags.kt
@@ -8,4 +8,5 @@ import site.siredvin.turtlematic.Turtlematic
 object BlockTags {
     val CAPTURE_BLACKLIST = TagKey.create(Registry.BLOCK_REGISTRY, ResourceLocation(Turtlematic.MOD_ID, "capture_blacklist"))
     val ENCHANTMENT_POWER_PROVIDER = TagKey.create(Registry.BLOCK_REGISTRY, ResourceLocation(Turtlematic.MOD_ID, "enchantment_power_provider"))
+    val HUSBANDRY_EXTRA_CROPS = TagKey.create(Registry.BLOCK_REGISTRY, ResourceLocation(Turtlematic.MOD_ID, "husbandry_extra_crops"))
 }